A lot of Server-side and Client-side code
Server
Finished Login end-points
Started to work on Data end-points

Client
Finished UI for login and registration
working on Graphs and Management

+++ ServerCode/WEA_AspOneFilipJakab/Models/TransactionModel.cs
@@ -16,9 +16,9 @@ namespace WEA_AspOneFilipJakab.Models
 
 		public List<int> TagIds { get; set; }
 
-		public long? Longitude { get; set; }
+		public decimal? Longitude { get; set; }
 
-		public long? Latitude { get; set; }
+		public decimal? Latitude { get; set; }
 
 		public int? CategoryId { get; set; }
 	}
